fix clang-tidy errors 5

+++ tasks/klimenko_v_max_matrix_elems_val/mpi/src/ops_mpi.cpp
@@ -2,8 +2,8 @@
 
 #include <mpi.h>
 
+#include <algorithm>
 #include <climits>
-#include <cstddef>
 
 #include "klimenko_v_max_matrix_elems_val/common/include/common.hpp"
 
@@ -30,7 +30,7 @@ bool KlimenkoVMaxMatrixElemsValMPI::RunImpl() {
   MPI_Comm_rank(MPI_COMM_WORLD, &rank);
   MPI_Comm_size(MPI_COMM_WORLD, &size);
 
-  int n = matrix.size();
+  int n = static_cast<int>(matrix.size());
   MPI_Bcast(&n, 1, MPI_INT, 0, MPI_COMM_WORLD);
   if (n == 0) {
     if (rank == 0) {
